+++ NEWS.md
@@ -9,6 +9,8 @@
   * check labels now test for spaces rather than check dashes
   * `lint_bib` does not add commas after `@string` fields.
   * `inputs_of` does not emit arcane warning if multiple `\end{document}`s exist.
+  * `check_escapes` do not check content within `tikzpicture` environment
+  * `check_dashes` does not check display equation lines.
 
 # TeXCheckR 0.4.3
 * Use `hunspell`'s ignore option when a dictionary addition is not respected